requirements.txt
Click==7.0
dash==1.7.0
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
pandas==0.25.3
plotly==4.4.1
python-dateutil==2.8.1
pytz==2019.3
retrying==1.3.3
six==1.13.0
Werkzeug==0.16.0
google-api-python-client==1.6.5
oauth2client==4.1.2
google-cloud-bigquery==0.29.0